+++ .github/files/generate-ci-matrix.php
@@ -6,8 +6,6 @@
  * @package automattic/jetpack
  */
 
-// phpcs:disable WordPress.WP.GlobalVariablesOverride
-
 chdir( __DIR__ . '/../../' );
 
 // Default versions for PHP and Node.
@@ -118,6 +116,14 @@ $matrix[] = array(
 	'name'    => 'JS tests',
 	'script'  => 'test-js',
 	'timeout' => 15, // 2021-01-18: Successful runs seem to take ~5 minutes.
+);
+
+// Add Coverage tests.
+$matrix[] = array(
+	'name'    => 'Code coverage',
+	'script'  => 'test-coverage',
+	'wp'      => 'latest',
+	'timeout' => 40, // 2024-10-30: Successful runs seem to take ~30 minutes. We'll need to improve that.
 );
 
 // END matrix definitions.