+++ projects/packages/forms/tests/php/contact-form/test-class.contact-form.php
@@ -15,7 +15,7 @@ use WorDBless\Posts;
 /**
  * Test class for Contact_Form
  *
- * @covers Contact_Form
+ * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form
  */
 class WP_Test_Contact_Form extends BaseTestCase {
 
@@ -577,10 +577,10 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	 * Tests that 'grunion_delete_old_spam()' deletes an old post that is marked as spam.
 	 *
 	 * @author tonykova
-	 * @covers Util::grunion_delete_old_spam
+	 * @covers \Automattic\Jetpack\Forms\ContactForm\Util::grunion_delete_old_spam
 	 */
 	public function test_grunion_delete_old_spam_deletes_an_old_post_marked_as_spam() {
-		// grunion_Delete_old_spam performs direct DB queries which cannot be tested outisde of a working WP install.
+		// grunion_Delete_old_spam performs direct DB queries which cannot be tested outside of a working WP install.
 		$this->markTestSkipped();
 		// @phan-suppress-next-line PhanPluginUnreachableCode
 		$post_id = wp_insert_post(
@@ -599,7 +599,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	 * Tests that 'grunion_delete_old_spam' does not delete a new post that is marked as spam.
 	 *
 	 * @author tonykova
-	 * @covers ::grunion_delete_old_spam
+	 * @covers \Automattic\Jetpack\Forms\ContactForm\Util::grunion_delete_old_spam
 	 */
 	public function test_grunion_delete_old_spam_does_not_delete_a_new_post_marked_as_spam() {
 		$post_id = wp_insert_post(
@@ -619,7 +619,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	 * Tests that token is left intact when there is not matching field.
 	 *
 	 * @author tonykova
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 */
 	public function test_token_left_intact_when_no_matching_field() {
 		$plugin       = Contact_Form_Plugin::init();
@@ -635,7 +635,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	 * Tests that token is replaced with an empty string when there is not value in field.
 	 *
 	 * @author tonykova
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 */
 	public function test_replaced_with_empty_string_when_no_value_in_field() {
 		$plugin       = Contact_Form_Plugin::init();
@@ -651,7 +651,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	 * Tests that token in curly brackets is replaced with the value when the name has whitespace.
 	 *
 	 * @author tonykova
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 */
 	public function test_token_can_replace_entire_subject_with_token_field_whose_name_has_whitespace() {
 		$plugin       = Contact_Form_Plugin::init();
@@ -667,7 +667,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	 * Tests that token with curly brackets is replaced with value.
 	 *
 	 * @author tonykova
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 */
 	public function test_token_with_curly_brackets_can_be_replaced() {
 		$plugin       = Contact_Form_Plugin::init();
@@ -715,7 +715,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Tests shortcode with commas and brackets.
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_array_values_with_commas_and_brackets() {
 		$shortcode = "[contact-field type='radio' options='\"foo\",bar&#044; baz,&#091;b&#092;rackets&#093;' label='fun &#093;&#091; times'/]";
@@ -726,7 +726,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Tests Gutenblock input with commas and brackets.
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_array_values_with_commas_and_brackets_from_gutenblock() {
 		$attr = array(
@@ -741,7 +741,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for text field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_text_field_renders_as_expected() {
 		$attributes = array(
@@ -760,7 +760,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for email field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_email_field_renders_as_expected() {
 		$attributes = array(
@@ -779,7 +779,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for url field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_url_field_renders_as_expected() {
 		$attributes = array(
@@ -798,7 +798,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for telephone field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_telephone_field_renders_as_expected() {
 		$attributes = array(
@@ -817,7 +817,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for date field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_date_field_renders_as_expected() {
 		$attributes = array(
@@ -837,7 +837,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for textarea field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_textarea_field_renders_as_expected() {
 		$attributes = array(
@@ -856,7 +856,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for checkbox field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_checkbox_field_renders_as_expected() {
 		$attributes = array(
@@ -875,7 +875,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Multiple fields
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_checkbox_multiple_field_renders_as_expected() {
 		$attributes = array(
@@ -895,7 +895,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for radio field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_radio_field_renders_as_expected() {
 		$attributes = array(
@@ -915,7 +915,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test for select field_renders
 	 *
-	 * @covers Contact_Form_Field
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Field
 	 */
 	public function test_make_sure_select_field_renders_as_expected() {
 		$attributes = array(
@@ -1208,7 +1208,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test get_export_data_for_posts with fully vaid data input.
 	 *
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 * @group csvexport
 	 */
 	public function test_get_export_data_for_posts_fully_valid_data() {
@@ -1328,7 +1328,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test get_export_data_for_posts with single invalid entry for post meta
 	 *
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 * @group csvexport
 	 */
 	public function test_get_export_data_for_posts_invalid_single_entry_meta() {
@@ -1433,7 +1433,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test get_export_data_for_posts with invalid all entries for post meta
 	 *
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 * @group csvexport
 	 */
 	public function test_get_export_data_for_posts_invalid_all_entries_meta() {
@@ -1523,7 +1523,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test get_export_data_for_posts with single invalid entry for parsed fields.
 	 *
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 * @group csvexport
 	 */
 	public function test_get_export_data_for_posts_single_invalid_entry_for_parse_fields() {
@@ -1636,7 +1636,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test get_export_data_for_posts with all entries for parsed fields invalid.
 	 *
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 * @group csvexport
 	 */
 	public function test_get_export_data_for_posts_all_entries_for_parse_fields_invalid() {
@@ -1679,7 +1679,7 @@ class WP_Test_Contact_Form extends BaseTestCase {
 	/**
 	 * Test map_parsed_field_contents_of_post_to_field_names
 	 *
-	 * @covers Contact_Form_Plugin
+	 * @covers Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin
 	 * @group csvexport
 	 */
 	public function test_map_parsed_field_contents_of_post_to_field_names() {