+++ projects/plugins/boost/app/assets/src/js/features/ui/upgraded/upgraded.tsx
@@ -1,8 +1,6 @@
 import { __ } from '@wordpress/i18n';
-import styles from './upgraded.module.scss';
+import Pill from '$features/ui/pill/pill';
 
-const Upgraded = () => (
-	<span className={ styles.badge }>{ __( 'Upgraded', 'jetpack-boost' ) }</span>
-);
+const Upgraded = () => <Pill text={ __( 'Upgraded', 'jetpack-boost' ) } altVersion={ true } />;
 
 export default Upgraded;