+++ projects/packages/sync/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.14.4] - 2024-11-04
+### Added
+- Enable test coverage. [#39961]
+
 ## [3.14.3] - 2024-10-25
 ### Fixed
 - Hooks: Hook init_sync_cron_jobs into init to ensure translation loading within the function is not triggered too early. [#39841]
@@ -1322,6 +1326,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Move sync to a classmapped package
 
+[3.14.4]: https://github.com/Automattic/jetpack-sync/compare/v3.14.3...v3.14.4
 [3.14.3]: https://github.com/Automattic/jetpack-sync/compare/v3.14.2...v3.14.3
 [3.14.2]: https://github.com/Automattic/jetpack-sync/compare/v3.14.1...v3.14.2
 [3.14.1]: https://github.com/Automattic/jetpack-sync/compare/v3.14.0...v3.14.1